Initial commit
Setup Cmake with LLVM and ANTLR (Kinda)

Also includes a parser that generates c++ classes.
Now I can start looking at LLVM

Antlr is weird and the easiest way to do it so far is to
download a java archive to use. Stolen from various
tutorials on the net, and not really optimal. Hopefully
will update later

LLVM was nice to set up